backend/requirements.txt
fastapi
uvicorn
pydantic
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dotenv
requests
openai
